Rename api folders to routes in user-service
